Moves some logic from data/stellar_fluxes.py to sn_analysis/reference.py
